Rename the CleanseListService, call it from the the main lambda
